refactor(provider): move passport catalog use case to application layer
